Consolidate keyvault options and model files into single headers.
