--- sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/deleted_key.hpp
@@ -1,56 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// SPDX-License-Identifier: MIT
-
-/**
- * @file
- * @brief Represents a Key Vault key that has been deleted, allowing it to be recovered, if needed.
- *
- */
-
-#pragma once
-
-#include <azure/core/datetime.hpp>
-
-#include "azure/keyvault/keys/keyvault_key.hpp"
-
-namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
-
-  /**
-   * @brief Represents a Key Vault key that has been deleted, allowing it to be recovered, if
-   * needed.
-   *
-   */
-  struct DeletedKey final : public KeyVaultKey
-  {
-    /**
-     * @brief A recovery URL that can be used to recover it.
-     *
-     */
-    std::string RecoveryId;
-
-    /**
-     * @brief Construct an empty DeletedKey
-     *
-     */
-    DeletedKey() = default;
-
-    /**
-     * @brief Construct a new Deleted Key object.
-     *
-     * @param name The name of the deleted key.
-     */
-    DeletedKey(std::string name) : KeyVaultKey(name) {}
-
-    /**
-     * @brief Indicate when the key was deleted.
-     *
-     */
-    Azure::DateTime DeletedDate;
-
-    /**
-     * @brief Indicate when the deleted key will be purged.
-     *
-     */
-    Azure::DateTime ScheduledPurgeDate;
-  };
-}}}} // namespace Azure::Security::KeyVault::Keys
