+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/recover_deleted_key_operation.hpp
@@ -15,7 +15,7 @@
 #include <azure/core/operation_status.hpp>
 #include <azure/core/response.hpp>
 
-#include "azure/keyvault/keys/keyvault_key.hpp"
+#include "azure/keyvault/keys/key_client_models.hpp"
 
 #include <memory>
 #include <string>
