+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keyvault_keys.hpp
@@ -22,17 +22,9 @@
 #include "azure/keyvault/keys/cryptography/verify_result.hpp"
 #include "azure/keyvault/keys/cryptography/wrap_result.hpp"
 #include "azure/keyvault/keys/delete_key_operation.hpp"
-#include "azure/keyvault/keys/deleted_key.hpp"
 #include "azure/keyvault/keys/dll_import_export.hpp"
-#include "azure/keyvault/keys/import_key_options.hpp"
-#include "azure/keyvault/keys/json_web_key.hpp"
 #include "azure/keyvault/keys/key_client.hpp"
 #include "azure/keyvault/keys/key_client_options.hpp"
-#include "azure/keyvault/keys/key_create_options.hpp"
-#include "azure/keyvault/keys/key_curve_name.hpp"
-#include "azure/keyvault/keys/key_operation.hpp"
-#include "azure/keyvault/keys/key_properties.hpp"
-#include "azure/keyvault/keys/key_type.hpp"
-#include "azure/keyvault/keys/keyvault_key.hpp"
+#include "azure/keyvault/keys/key_client_models.hpp"
 #include "azure/keyvault/keys/list_keys_responses.hpp"
 #include "azure/keyvault/keys/recover_deleted_key_operation.hpp"
