+++ sdk/keyvault/azure-security-keyvault-keys/src/private/local_cryptography_provider.hpp
@@ -11,7 +11,7 @@
 
 #include "cryptography_provider.hpp"
 
-#include "azure/keyvault/keys/keyvault_key.hpp"
+#include "azure/keyvault/keys/key_client_models.hpp"
 
 #include <memory>
 #include <string>
